+++ Provider/src/test/java/org/example/mapper/CalculateTest.java
@@ -8,7 +8,7 @@ import java.math.BigDecimal;
 import static org.assertj.core.api.Assertions.assertThat;
 
 class CalculateTest {
-    private final double income = 10_505.00;
+    private final BigDecimal income = BigDecimal.valueOf(10_505.00);
     private final double percentageTaxRate = 10.00;
 
     @Test
